+++ @xen-orchestra/lite/src/stories/web-core/ui/link/ui-link.story.vue
@@ -9,8 +9,8 @@
       prop('isPrimary').bool().widget(),
       prop('primaryTooltip').str().widget(),
       prop('disabled').bool().widget(),
-      prop('disabled').bool().widget(),
       prop('target').enum('_blank', '_self').widget(),
+      prop('wrap').bool().widget().help('Control whether the link wraps. If not, the text has ellipsis.'),
     ]"
   >
     <UiLink v-bind="properties">This is a link</UiLink>