+++ @xen-orchestra/web/src/modules/pool/components/list/panel/PoolSidePanel.vue
@@ -30,6 +30,7 @@
                 icon="object:pool"
                 size="small"
                 :to="{ name: '/pool/[id]/dashboard', params: { id: server.poolId } }"
+                wrap
               >
                 {{ server.poolNameLabel }}
               </UiLink>
@@ -90,6 +91,7 @@
               :to="{ name: '/host/[id]/dashboard', params: { id: primaryHost.id } }"
               is-primary
               :primary-tooltip="t('master')"
+              wrap
             >
               {{ primaryHost.name_label }}
             </UiLink>
@@ -155,6 +157,7 @@
             :to="{ name: '/host/[id]/dashboard', params: { id: host.id } }"
             :icon="`object:host:${toLower(host.power_state)}`"
             size="small"
+            wrap
           >
             {{ host.name_label }}
             <VtsIcon v-if="primaryHost?.id === host.id" accent="info" name="status:primary-circle" size="medium" />