+++ @xen-orchestra/web/src/modules/snapshot/components/list/panel/card-items/SnapshotVdiLinkCardItem.vue
@@ -1,5 +1,7 @@
 <template>
-  <UiLink :icon="vbdsSnapshotStatus" size="medium" :href="xo5VmSnapshotVdiHref">{{ vdi.name_label }}</UiLink>
+  <UiLink :icon="vbdsSnapshotStatus" size="medium" :href="xo5VmSnapshotVdiHref" wrap>
+    {{ vdi.name_label }}
+  </UiLink>
 </template>
 
 <script setup lang="ts">